Implement fast local execution when local worker has cores available

+++ UET/Redpoint.OpenGE.Component.Worker/DefaultBlobManager.cs
@@ -76,32 +76,35 @@
                     var targetPath = ConvertAbsolutePathToBuildDirectoryPath(
                         targetDirectory,
                         kv.Key);
-                    Directory.CreateDirectory(Path.GetDirectoryName(targetPath)!);
-                    if (virtualised.Contains(kv.Key))
+                    if (!File.Exists(targetPath))
                     {
-                        // Grab the blob content, replace {__OPENGE_VIRTUAL_ROOT__} and then emit it.
-                        using var sourceStream = new FileStream(
-                            Path.Combine(blobsPath, kv.Value.HexString()),
-                            FileMode.Open,
-                            FileAccess.Read,
-                            FileShare.Read);
-                        using var reader = new StreamReader(sourceStream, leaveOpen: true);
-                        var content = await reader.ReadToEndAsync(cancellationToken);
-                        content = content.Replace("{__OPENGE_VIRTUAL_ROOT__}", shortenedTargetDirectory);
-                        using var targetStream = new FileStream(
-                            targetPath,
-                            FileMode.Create,
-                            FileAccess.Write,
-                            FileShare.None);
-                        using var writer = new StreamWriter(targetStream, leaveOpen: true);
-                        await writer.WriteAsync(content);
-                    }
-                    else
-                    {
-                        File.Copy(
-                            Path.Combine(blobsPath, kv.Value.HexString()),
-                            targetPath,
-                            true);
+                        Directory.CreateDirectory(Path.GetDirectoryName(targetPath)!);
+                        if (virtualised.Contains(kv.Key))
+                        {
+                            // Grab the blob content, replace {__OPENGE_VIRTUAL_ROOT__} and then emit it.
+                            using var sourceStream = new FileStream(
+                                Path.Combine(blobsPath, kv.Value.HexString()),
+                                FileMode.Open,
+                                FileAccess.Read,
+                                FileShare.Read);
+                            using var reader = new StreamReader(sourceStream, leaveOpen: true);
+                            var content = await reader.ReadToEndAsync(cancellationToken);
+                            content = content.Replace("{__OPENGE_VIRTUAL_ROOT__}", shortenedTargetDirectory);
+                            using var targetStream = new FileStream(
+                                targetPath,
+                                FileMode.Create,
+                                FileAccess.Write,
+                                FileShare.None);
+                            using var writer = new StreamWriter(targetStream, leaveOpen: true);
+                            await writer.WriteAsync(content);
+                        }
+                        else
+                        {
+                            File.Copy(
+                                Path.Combine(blobsPath, kv.Value.HexString()),
+                                targetPath,
+                                true);
+                        }
                     }
                 });
         }
